implement more of the CDC; BIOS now gets to the point of initiating a data transfer
